will be using only cblas in lorenzo_bbgky. No other dependencies

+++ lorenzo_bbgky/lorenzo_bbgky.h
@@ -2,10 +2,7 @@
 #include <math.h>
 #include <stdlib.h>
 #include <limits.h>
-#include <gsl/gsl_vector.h>
-#include <gsl/gsl_matrix.h>
-#include <gsl/gsl_blas.h>
+#include <cblas.h>
 
 /* Functions */
-int dsdg (double *, double *, double *, double *, double, double, double,
-	  double *);
+int dsdg (double *, double *, double *, double *, double, int, double, double *);
